requirements.txt
asgiref==3.7.2
decouple==0.0.7
Django==5.0
djangorestframework==3.14.0
dotenv==0.0.5
Jinja2==3.1.2
MarkupSafe==2.1.3
Pillow==10.1.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
